financial-agent-py/requirements.txt
fastapi>=0.115.2
uvicorn>=0.30.0
openai>=1.3.7
python-dotenv>=1.0.0
pydantic>=2.8.2
httpx>=0.28.1
payments-py>=0.8.0